\name{simulated_data}
\alias{simulated_data}
\docType{data}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Simulated data from Simulation 1 in cited paper.
}
\description{
Simulated data with two groups, used to illustrate \code{\link{cdghmm}}
}
\usage{
data(simulated_data)
}
%- maybe also 'usage' for other objects documented here.
\source{
These data were simulated using \code{R}.
}
%\references{
%}
\keyword{datasets}
